requirements.txt
ultralytics
requests
opencv-python
azure-storage-blob
geopy
numpy
geocoder